chore: finished cleanup (moved revit render material proxy cache out)
