+++ Converters/Revit/Speckle.Converters.RevitShared.DependencyInjection/RevitConverterModule.cs
@@ -22,7 +22,7 @@ public class RevitConverterModule : ISpeckleModule
     builder.AddScoped<IRootToHostConverter, RevitRootToHostConverter>();
     builder.AddSingleton(new RevitContext());
 
-    builder.AddSingleton(new RenderMaterialProxyCacheSingleton());
+    builder.AddSingleton(new RevitRenderMaterialProxyCacheSingleton());
 
     // POC: do we need ToSpeckleScalingService as is, do we need to interface it out?
     builder.AddScoped<ScalingServiceToSpeckle>();
